server/vps/requirements.txt
# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# ── Telegram ──────────────────────────────────────────────────────────────────
telethon>=1.37.0

# ── Database ──────────────────────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Cifratura credenziali MT5 ─────────────────────────────────────────────────
cryptography>=43.0.0

# ── MetaTrader 5 (solo Windows) ───────────────────────────────────────────────
# MetaTrader5>=5.0.4424
